requirements_production.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
tensorflow==2.13.0
tensorflow-serving-api==2.13.0
pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
scipy==1.11.2

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.4.2

# Translation
googletrans==4.0.0-rc1
transformers==4.35.0
torch==2.0.1
azure-cognitiveservices-speech==1.32.1
gtts==2.4.0

# Database and Cache
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
pyjwt==2.8.0
cryptography==41.0.5
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.18.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
psutil==5.9.6
gputil==1.4.0

# Cloud Services
boto3==1.28.84
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Data Processing
pandas==2.1.3
h5py==3.10.0
pyyaml==6.0.1

# Model Optimization
tensorflow-model-optimization==0.7.5
onnx==1.15.0
onnxruntime-gpu==1.16.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0